+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
@@ -90,35 +89,6 @@ func init() {
 	if e := os.Getenv("ACCESS_LOG"); e == "true" {
 		accessLog = true
 	}
-}
-
-func handler(w http.ResponseWriter, req *http.Request) {
-	if accessLog {
-		log.Printf("%s %s %s", req.RemoteAddr, req.Method, req.RequestURI)
-	}
-	time.Sleep(responseSleep)
-	_, _ = w.Write(responseBody)
-}
-
-func echoHandler(w http.ResponseWriter, req *http.Request) {
-	if accessLog {
-		log.Printf("%s %s %s", req.RemoteAddr, req.Method, req.RequestURI)
-	}
-	respMap := map[string]interface{}{
-		"Header":     req.Header,
-		"Form":       req.Form,
-		"Proto":      req.Proto,
-		"Method":     req.Method,
-		"Host":       req.Host,
-		"RequestURI": req.RequestURI,
-		"RemoteAddr": req.RemoteAddr,
-	}
-	resp, err := json.MarshalIndent(respMap, "", "    ")
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	_, _ = w.Write(resp)
 }
 
 func main() {
